add the `url` universal shortcode
i needed URLs to be environment-dependent because this content is not stored in the public root of my site, but the dev server doesn't reflect that.

problem i found is that the environment is not passed into the templates (and i'm not really fond of passing the contents of `.env` into templates), nor is the contents of `site.json` available to shortcodes (only `page` data is).

my solution was to switch `site.json` to a JS export, which 11ty can handle, but also `require` it into `.eleventy.js` so the shortcode could access it.

this also requires i switch the template engine in each `.html` file, so that i have access to both `set` and `~`-based concatenation.